requirements.dev.txt
pycryptodome
pytest
pytest-cov
wheel
setuptools
twine

